+++ .github/copy_artifacts.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 mkdir plug-a-thon-mdibs
-cp -a generated_mdibs/. plug-a-thon-mdibs/
+cp -a build/mdib/build/generated_mdibs/. plug-a-thon-mdibs/
+cp resources/test-sequence.json plug-a-thon-mdibs/