+++ CHANGELOG.md
@@ -1,12 +1,18 @@
 # Changelog
 
-All notable changes to the MDIB will be documented in this file.
+All notable changes to the MDIB and test specification will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Each section shall contain a list of action items of the following format: `<brief one-sentence description of what has been done> (#\[<issue number]\(<URL>)).`
 
 ## [Unreleased]
+
+## [2.4.0] - 2026-02-13
+
+### Changed
+
+- Test specification is now provided as a JSON as part of the release download
 
 ### Added